# server/swagger.yml
openapi: 3.1.0
info:
  title: PLANT System API
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8000
    description: 本機開發環境
security: []
...
paths:
...
  /api/projects/{project_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Projects]
      summary: 取得專案 Artifact
      operationId: getProject
      security:
        - ActivationCookie: []
        - {}
      responses:
        "200":
          description: 專案資料
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Projects]
      summary: 刪除專案
      operationId: deleteProject
      security:
        - ActivationCookie: []
      responses:
        "200":
          description: 專案刪除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteProjectResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 專案正在執行 Run，或有附件、刪除等專案操作正在進行，無法刪除
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                detail: Project is running or another project operation is in progress
...
  /api/projects/{project_id}/artifacts:
    get:
      tags: [Artifacts]
      summary: 取得專案檔案樹
      operationId: getProjectArtifactTree
      security:
        - ActivationCookie: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: 檔案樹
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactTreeResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/projects/{project_id}/files:
    get:
      tags: [Artifacts]
      summary: 讀取專案檔案
      operationId: readProjectFile
      security:
        - ActivationCookie: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: path
          in: query
          required: true
          description: 相對於專案目錄的檔案路徑
          schema:
            type: string
          example: artifact/requirements.json
      responses:
        "200":
          description: 檔案內容或 Base64 圖片
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /{project_id}/manual:
    get:
      tags: [Public Manual]
      summary: 取得專案 Manual 首頁
      operationId: getProjectManual
      security: [{ ActivationCookie: [] }, {}]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200": { $ref: "#/components/responses/HtmlFile" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /{project_id}/manual/:
    get:
      tags: [Public Manual]
      summary: 取得專案 Manual 首頁（尾斜線）
      operationId: getProjectManualWithTrailingSlash
      security: [{ ActivationCookie: [] }, {}]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200": { $ref: "#/components/responses/HtmlFile" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /{project_id}/manual.zip:
    get:
      tags: [Public Manual]
      summary: 下載專案離線 Manual
      operationId: downloadProjectManualZip
      security: [{ ActivationCookie: [] }, {}]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200": { $ref: "#/components/responses/ZipFile" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/projects/{project_id}/manual.zip:
    get:
      tags: [Artifacts]
      summary: 透過 API 下載專案離線 Manual
      operationId: downloadProjectManualZipApi
      security: [{ ActivationCookie: [] }, {}]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200": { $ref: "#/components/responses/ZipFile" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /manual.zip:
    get:
      tags: [Public Manual]
      summary: 下載共用離線 Manual
      operationId: downloadSharedManualZip
      responses:
        "200": { $ref: "#/components/responses/ZipFile" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/manual.zip:
    get:
      tags: [Artifacts]
      summary: 透過 API 下載共用離線 Manual
      operationId: downloadSharedManualZipApi
      responses:
        "200": { $ref: "#/components/responses/ZipFile" }
        "404": { $ref: "#/components/responses/NotFound" }
  /{project_id}/manual/srs:
    get:
      tags: [Public Manual]
      summary: 取得 SRS HTML
      operationId: getProjectSrs
      security: [{ ActivationCookie: [] }, {}]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200": { $ref: "#/components/responses/HtmlFile" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /{project_id}/manual/dr:
    get:
      tags: [Public Manual]
      summary: 取得 Design Rationale HTML
      operationId: getProjectDesignRationale
      security: [{ ActivationCookie: [] }, {}]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200": { $ref: "#/components/responses/HtmlFile" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /{project_id}/manual/models/{file_path}:
    get:
      tags: [Public Manual]
      summary: 取得 Manual 模型檔案
      operationId: getProjectManualModelFile
      security: [{ ActivationCookie: [] }, {}]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FilePath"
      responses:
        "200": { $ref: "#/components/responses/BinaryFile" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /{project_id}/manual/results/{file_path}:
    get:
      tags: [Public Manual]
      summary: 取得 Manual results 檔案
      operationId: getProjectManualResultFile
      security: [{ ActivationCookie: [] }, {}]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FilePath"
      responses:
        "200": { $ref: "#/components/responses/BinaryFile" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /{project_id}/manual/artifact/{file_path}:
    get:
      tags: [Public Manual]
      summary: 取得 Manual artifact 檔案
      operationId: getProjectManualArtifactFile
      security: [{ ActivationCookie: [] }, {}]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FilePath"
      responses:
        "200": { $ref: "#/components/responses/BinaryFile" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /{project_id}/manual/output/{file_path}:
    get:
      tags: [Public Manual]
      summary: 取得 Manual output 檔案
      operationId: getProjectManualOutputFile
      security: [{ ActivationCookie: [] }, {}]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FilePath"
      responses:
        "200": { $ref: "#/components/responses/BinaryFile" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /{project_id}/manual/agent/{file_path}:
    get:
      tags: [Public Manual]
      summary: 取得 Manual Agent 共用資源
      operationId: getProjectManualAgentFile
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FilePath"
      responses:
        "200": { $ref: "#/components/responses/BinaryFile" }
        "404": { $ref: "#/components/responses/NotFound" }
  /{project_id}/manual/{file_path}:
    get:
      tags: [Public Manual]
      summary: 取得專案 Manual 檔案
      operationId: getProjectManualFile
      security: [{ ActivationCookie: [] }, {}]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FilePath"
      responses:
        "200": { $ref: "#/components/responses/BinaryFile" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /manual/{file_path}:
    get:
      tags: [Public Manual]
      summary: 取得共用 Manual 資源
      operationId: getSharedManualFile
      parameters:
        - $ref: "#/components/parameters/FilePath"
      responses:
        "200": { $ref: "#/components/responses/BinaryFile" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    ActivationCookie:
      type: apiKey
      in: cookie
      name: plant_activation
      description: 透過 POST /api/secrets/activation-code 取得的 HttpOnly Cookie
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema: { type: string }
      example: "173949223214"
...
    FilePath:
      name: file_path
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: 請求資料或操作條件不合法
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: 缺少啟動權限或專案目前不可公開讀取
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example: { detail: 需要啟動碼才能執行此操作 }
    NotFound:
      description: 找不到指定資源
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    BinaryFile:
      description: 檔案內容；實際 Content-Type 依副檔名決定
      content:
        application/octet-stream:
          schema: { type: string, format: binary }
    HtmlFile:
      description: HTML 文件
      content:
        text/html:
          schema: { type: string }
    ZipFile:
      description: ZIP 壓縮檔
      content:
        application/zip:
          schema: { type: string, format: binary }
  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          description: 錯誤訊息或詳細資料
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
...
    ProjectResponse:
      type: object
      required: [project_id, project]
      properties:
        project_id: { type: string }
        project:
          type: object
          additionalProperties: true
    DeleteProjectResponse:
      type: object
      required: [project_id, deleted]
      properties:
        project_id: { type: string }
        deleted: { type: boolean, const: true }
...
    FileTreeNode:
      type: object
      required: [path, name, kind]
      properties:
        path: { type: string }
        name: { type: string }
        kind: { type: string, enum: [file, directory] }
        size: { type: [integer, "null"] }
        editable: { type: boolean }
        previewable: { type: boolean }
    ArtifactTreeResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/FileTreeNode" }
    FileContent:
      type: object
      required: [path, type, encoding, content, editable, readonly]
      properties:
        path: { type: string }
        type: { type: string }
        encoding: { type: string, enum: [utf-8, base64] }
        content: { type: string }
        editable: { type: boolean }
        readonly: { type: boolean }
        mime: { type: string }
...